Shiny now determines port, dependency update
Shiny now chooses the port it runs at, not Electron, avoiding a firewall prompt. Electron listens for the _Listening on http://..._ message from Shiny, then connects, instead of repeatedly trying until Shiny is running, decreasing startup time and generating less clutter in the log. The argument to allow Shiny to run at a fixed port is removed, an argument is added to pass arguments to the function starting your app, which, among others, allows you to set a fixed port. Dependencies are updated.